remove unncessary file
